% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{client}
\alias{client}
\title{Create a client connection to the Zendesk API}
\usage{
client(url = Sys.getenv("ZENDESK_URL"), user = Sys.getenv("ZENDESK_USER"),
  token = Sys.getenv("ZENDESK_TOKEN"), debugging = FALSE)
}
\arguments{
\item{url}{The base URL to the Zendesk API service}

\item{user}{The user to access Zendesk}

\item{token}{= The token to access Zendesk}

\item{debugging}{set to TRUE to increase stdout logging and provide additional debugging detail}
}
\description{
Create a client connection to the Zendesk API
}
\examples{
client(url = 'https://organization.zendesk.com/api/v2',
    user = 'zendesk_user',
    token = 'zendesk_token')

}
