% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ticket.R
\name{get_ticket}
\alias{get_ticket}
\title{Create a client connection to the Zendesk API to retreive a Zendesk ticket}
\usage{
get_ticket(ticket, cli = client())
}
\arguments{
\item{ticket}{a unique string to identify specific Zendesk ticket to get from API}

\item{cli}{the Zendesk client [created from ::client]}
}
\value{
object with three attributes:
    content - json response from API parsed into named list
    headers - headers passed with API request for ticket
    response - API response to request
}
\description{
Get a Zendesk ticket from API call by passing Zendesk ticket number
}
\examples{
get_ticket('123455', cli = client())

}
