% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_ticket.R
\name{post_ticket}
\alias{post_ticket}
\title{Create a client connection to the Zendesk API}
\usage{
post_ticket(requester_name, requester_email, body, subject, group,
  tags = list(), public, cli = client())
}
\arguments{
\item{requester_name}{a string to identify customer who should be listed as ticket requester (optional, if NULL the email will pull in corresponding name)}

\item{requester_email}{a string to identify customer who should be listed as ticket requester}

\item{body}{a character string that is passed to the body of the Zendesk ticket}

\item{subject}{the subject of your Zendesk ticket}

\item{group}{an integer to indicate the group_id for your new ticket (optional, can be set to NULL)}

\item{tags}{a list of character strings to set as the ticket tags for your new ticket (optional, can be set to NULL)}

\item{public}{an optional logical value to indicate if the ticket comment should be public (TRUE) or privet (FALSE), default value FALSE}

\item{cli}{the Zendesk client [created from ::client]}
}
\value{
Zendesk ticket number for successfully created ticket
}
\description{
Allows user to post a new ticket to the Zendesk API on behalf of a customer
}
\examples{
post_ticket(
  requester_name = 'fake_name',
  requester_email = 'fake_email',
  body = 'comment in body of ticket',
  subject = 'subject of ticket',
  group = 1234567,
  tags = list(),
  public = FALSE
)

}
