% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zendesk.R
\name{zdExtractData}
\alias{zdExtractData}
\title{Extracts data from list of tickets as parsed from json by httr::content}
\usage{
zdExtractData(items, fields)
}
\arguments{
\item{items}{unnamed list of zendesk objects with attributes}

\item{fields}{string vector of fields to be extracted from the object}
}
\value{
data.table with relevant object attributes
}
\description{
Extracts data from list of tickets as parsed from json by httr::content
}
