% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zendesk.R
\name{zdGetObjects}
\alias{zdGetObjects}
\title{Gets Zendesk objects for a given subdomain as a data.table}
\usage{
zdGetObjects(type, subdomain, start.time)
}
\arguments{
\item{type}{type of the object to be returned (users, tickets)}

\item{subdomain}{organisation subdomain on zendesk}

\item{start.time}{starting point for new objects based on the update_date}
}
\value{
list of zendesk objects according to the type
}
\description{
Gets Zendesk objects for a given subdomain as a data.table
}
