% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zendesk.R
\name{zdGetPath}
\alias{zdGetPath}
\title{Makes url for incremental object extraction from Zendesk}
\usage{
zdGetPath(type, start.time)
}
\arguments{
\item{type}{type of the object to be returned (users, tickets)}

\item{start.time}{starting point for new objects based on the update_date}
}
\description{
Makes url for incremental object extraction from Zendesk
}
