% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zendesk.R
\name{zdGetTickets}
\alias{zdGetTickets}
\title{Gets Zendesk tickets via increment API}
\usage{
zdGetTickets(subdomain, start.time)
}
\arguments{
\item{subdomain}{organisation subdomain on zendesk}

\item{start.time}{starting point for new tickets based on the update_date}
}
\value{
data.table with zendesk tickets
}
\description{
Gets Zendesk tickets via increment API
}
