% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zendesk.R
\name{zdGetUsers}
\alias{zdGetUsers}
\title{Gets Zendesk users via increment API}
\usage{
zdGetUsers(subdomain, start.time)
}
\arguments{
\item{subdomain}{organisation subdomain on zendesk}

\item{start.time}{starting point for new users based on the update_date}
}
\value{
data.table with zendesk users
}
\description{
Gets Zendesk users via increment API
}
