% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zendesk.R
\name{zdProcessResponse}
\alias{zdProcessResponse}
\title{Processes http response and returns a list of results}
\usage{
zdProcessResponse(response)
}
\arguments{
\item{response}{http response from the Zendesk API}
}
\value{
list of resutls parsed from the response
}
\description{
Processes http response and returns a list of results
}
