% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\docType{class}
\name{ZendeskClient-class}
\alias{ZendeskClient-class}
\title{An S4 class to represent a client connection}
\description{
An S4 class to represent a client connection
}
\section{Slots}{

\describe{
\item{\code{user}}{the Zendesk user (for use with API tokens please add '/token' to the end of your username)}

\item{\code{token}}{the user's Zendesk API token}

\item{\code{base_url}}{the url for c}
}}

