% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_ticket.R
\name{post_ticket}
\alias{post_ticket}
\title{Create a client connection to the Zendesk API}
\usage{
post_ticket(requester_name, requester_email, body, subject, tags = list(),
  cli = client())
}
\arguments{
\item{requester_name}{a string to identify customer who should be listed as ticket requester, with the ticket pointing to their account}

\item{requester_email}{a string to identify customer who should be listed as ticket requester}

\item{body}{a character string that is passed to the body of the Zendesk ticket}

\item{subject}{the subject of your Zendesk ticket}

\item{tags}{a list of custom tags for the new ticket}

\item{cli}{the Zendesk client [created from ::client]}
}
\value{
Zendesk ticket number for successfully created ticket
}
\description{
Allows user to post a new ticket to the Zendesk API on behalf of a customer
}
\examples{
post_ticket(
  requester_name = 'fake_name',
  requester_email = 'fake_email',
  body = 'comment in body of ticket',
  subject = 'subject of ticket'
)

}
